Move classes declared in GenericConsole all to their own files
