+++ src/games/strategy/debug/GenericConsole.java
@@ -98,80 +98,3 @@ public abstract class GenericConsole extends JFrame {
 }
 
 
-class ThreadReader implements Runnable {
-  private static final int CONSOLE_UPDATE_INTERVAL_MS = 100;
-  private final JTextArea m_text;
-  private final SynchedByteArrayOutputStream m_in;
-  private final boolean m_displayConsoleOnWrite;
-  private final GenericConsole parentConsole;
-
-  ThreadReader(final SynchedByteArrayOutputStream in, final JTextArea text, final boolean displayConsoleOnWrite,
-      GenericConsole parentConsole) {
-    m_in = in;
-    m_text = text;
-    m_displayConsoleOnWrite = displayConsoleOnWrite;
-    this.parentConsole = parentConsole;
-  }
-
-  @Override
-  public void run() {
-    while (true) {
-      m_text.append(m_in.readFully());
-      if (m_displayConsoleOnWrite && !parentConsole.isVisible()) {
-        parentConsole.setVisible(true);
-      }
-      ThreadUtil.sleep(CONSOLE_UPDATE_INTERVAL_MS);
-    }
-  }
-}
-
-
-/**
- * Allows data written to a byte output stream to be read
- * safely friom a seperate thread.
- * Only readFully() is currently threadSafe for reading.
- */
-class SynchedByteArrayOutputStream extends ByteArrayOutputStream {
-  private final Object lock = new Object();
-  private final PrintStream m_mirror;
-
-  SynchedByteArrayOutputStream(final PrintStream mirror) {
-    m_mirror = mirror;
-  }
-
-  public void write(final byte b) throws IOException {
-    synchronized (lock) {
-      m_mirror.write(b);
-      super.write(b);
-      lock.notifyAll();
-    }
-  }
-
-  @Override
-  public void write(final byte[] b, final int off, final int len) {
-    synchronized (lock) {
-      super.write(b, off, len);
-      m_mirror.write(b, off, len);
-      lock.notifyAll();
-    }
-  }
-
-  /**
-   * Read all data written to the stream.
-   * Blocks until data is available.
-   * This is currently the only threadsafe method for reading.
-   */
-  public String readFully() {
-    synchronized (lock) {
-      if (super.size() == 0) {
-        try {
-          lock.wait();
-        } catch (final InterruptedException ie) {
-        }
-      }
-      final String s = toString();
-      reset();
-      return s;
-    }
-  }
-}
